refactor: openShareDialog 함수로 분리
